fix: disable eslint in generated hooks, refactor package inter-dependencies
