+++ packages/schema/package.json
@@ -3,7 +3,7 @@
     "publisher": "zenstack",
     "displayName": "ZenStack Language Tools",
     "description": "A toolkit for building secure CRUD apps with Next.js + Typescript",
-    "version": "1.0.0-alpha.120",
+    "version": "1.0.0-alpha.121",
     "author": {
         "name": "ZenStack Team"
     },
