+++ packages/server/package.json
@@ -1,6 +1,6 @@
 {
     "name": "@zenstackhq/server",
-    "version": "1.0.0-alpha.120",
+    "version": "1.0.0-alpha.121",
     "displayName": "ZenStack Server-side Adapters",
     "description": "ZenStack server-side adapters",
     "homepage": "https://zenstack.dev",
@@ -25,7 +25,6 @@
     "dependencies": {
         "@zenstackhq/openapi": "workspace:*",
         "@zenstackhq/runtime": "workspace:*",
-        "@zenstackhq/sdk": "workspace:*",
         "change-case": "^4.1.2",
         "lower-case-first": "^2.0.2",
         "superjson": "^1.11.0",
