+++ packages/testtools/package.json
@@ -1,6 +1,6 @@
 {
   "name": "@zenstackhq/testtools",
-  "version": "1.0.0-alpha.120",
+  "version": "1.0.0-alpha.121",
   "description": "ZenStack Test Tools",
   "main": "index.js",
   "publishConfig": {
